% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{api_to_df}
\alias{api_to_df}
\title{Mine an API and save the results in a data frame}
\usage{
api_to_df(df, which_info, email = "")
}
\arguments{
\item{df}{Source data frame that contains DOI and/or ISSN columns}

\item{which_info}{What api will be mined? Either "doaj" or "upw" (unpaywall).}

\item{email}{email address of user (required for the Unpaywall API; otherwise can be left empty)}
}
\value{
a data frame
}
\description{
Using an API mining  function, query all unique
entries in a column, and return a data frame
with their results.
}
