% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{apply_custom}
\alias{apply_custom}
\title{Apply custom labels to a data frame}
\usage{
apply_custom(df, path)
}
\arguments{
\item{df}{Source data frame containing a system_id column}

\item{path}{The location of the custom file}
}
\value{
Data frame with an added column `custom_label`
}
\description{
Using a data file with columns with system IDs, this function
can be used to apply custom labels to publications.
Column headers are used as labels and applied to publications in
the data frame.
}
