% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{apply_doaj}
\alias{apply_doaj}
\title{Add DOAJ data to publications data frame}
\usage{
apply_doaj(df, doajdf, max_year = "previous")
}
\arguments{
\item{df}{Source data frame containing an ISSN column}

\item{doajdf}{Data frame resulting from DOAJ API mining (see `get_doaj()`)}

\item{max_year}{The journal must have been registered in the DOAJ before or during this year}
}
\value{
Data frame with a column `doaj` added
}
\description{
This function checks ISSNs in the publications data frame for their appearance in
the VSNU report, and adds a column called `doaj` containing True/False values
that indicate that a journal is present in the DOAJ.
}
