% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{apply_matches}
\alias{apply_matches}
\title{Add all VSNU, UPW, DOAJ matches to the data frame}
\usage{
apply_matches(df, doajdf, vsnudf, upwdf, max_year = "previous")
}
\arguments{
\item{df}{Source data frame containing doi and issn columns}

\item{doajdf}{Data frame resulting from DOAJ API mining (see `get_doaj()`)}

\item{vsnudf}{VSNU data (see `get_vsnu()`)}

\item{upwdf}{Data frame resulting from unpaywall API mining (see `get_upw()`)}

\item{max_year}{The journal must have been registered in the DOAJ before or during this year}
}
\value{
data frame with three extra columns: doaj, vsnu, upw
}
\description{
Combination function that applies all DOAJ, UPW, and VSNU data to the
data frame.
}
