% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{apply_upw}
\alias{apply_upw}
\title{Add Unpaywall data to publications data frame}
\usage{
apply_upw(df, upwdf)
}
\arguments{
\item{df}{Source data frame containing a doi column}

\item{upwdf}{Data frame resulting from unpaywall API mining (see `get_upw()`)}
}
\value{
Data frame with a column `upw` added
}
\description{
This function conflates the publications data frame with the results of mining
the Unpaywall API. The column containing the content of Unpaywall's `oa_color`
is added to the data frame and labeled `upw`.
}
