% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{apply_vsnu}
\alias{apply_vsnu}
\title{Add VSNU data to publications data frame}
\usage{
apply_vsnu(df, vsnudf)
}
\arguments{
\item{df}{Source data frame containing a doi column}

\item{vsnudf}{VSNU data (see `get_vsnu()`)}
}
\value{
Data frame with a column `vsnu` added
}
\description{
This function checks DOIs in the publications data frame for their appearance in
the VSNU report, and adds a column called `vsnu` containing True/False values
that indicate that publications were published under the VSNU deal.
}
