% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{check_info}
\alias{check_info}
\title{Screen for sufficient information}
\usage{
check_info(df, cutoff = 0.05, save = F)
}
\arguments{
\item{df}{The dataframe to check}

\item{cutoff}{The accepted proportion of "NONE" per organization unit}

\item{save}{Should the results be saved?}
}
\value{
Tibble with publications to check manually.
}
\description{
Screen the dataframe to see whether there is sufficient information
in each reporting unit to determine whether
}
