% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_doi}
\alias{clean_doi}
\title{Clean DOIs}
\usage{
clean_doi(column)
}
\arguments{
\item{column}{DOI(s)}
}
\value{
cleaned DOI(s), vectorized
}
\description{
This function removes url information, whitespace, and punctuation from DOIs
and returns them in correct format, all characters converted to lowercase. Any
additional DOIs separated by commas will also be removed so each field will contain
only a single DOI.
}
