% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_issn}
\alias{clean_issn}
\title{Clean ISSNs}
\usage{
clean_issn(column)
}
\arguments{
\item{column}{ISSN(s)}
}
\value{
cleaned ISSN(s), vectorized
}
\description{
This function removes whitespace and punctuation from ISSNs
and returns them in correct ISSN format.
An ISSN is formatted as NNNN-NNNX, where N is a number, and X can be any number or character.
If ISSNs are formatted differently, this can give errors when using a public API, so they can be
reformatted with this function.
A string with a different length will return NA.
}
