% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{custom_label}
\alias{custom_label}
\title{Generate the custom label column}
\usage{
custom_label(column, custom_list)
}
\arguments{
\item{column}{the column of system IDs that needs to be verified}

\item{custom_list}{the list generated by `get_custom`}
}
\value{
a vector with the length of `column` containing labels if applicable
}
\description{
Generate the custom label column
}
