% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{deduplicate}
\alias{deduplicate}
\title{Deduplication of publication entries}
\usage{
deduplicate(df)
}
\arguments{
\item{df}{The dataframe that needs to be deduplicated.}
}
\value{
The deduplicated dataframe.
}
\description{
Deduplication function that uses a single dataframe and applies a variety
of deduplication functions to subsets of the dataframe.
Required to use before reporting on a subset of publications, so that no bias is created
in the results if a publication is entered multiple times by different groups, eg.
}
