% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{deduplicate_method}
\alias{deduplicate_method}
\title{Deduplicate publication entries

Deduplication function that takes a dataframe and returns
the deduplicated dataframe based on either the combination of source file and
their internal ID in that source, or DOI.
Reports on the deduplication performed.}
\usage{
deduplicate_method(df, method)
}
\arguments{
\item{df}{The dataframe that needs to be deduplicate}

\item{method}{The deduplication method, can be either "doi" or "internal"}
}
\value{
The deduplicated dataframe
}
\description{
Deduplicate publication entries

Deduplication function that takes a dataframe and returns
the deduplicated dataframe based on either the combination of source file and
their internal ID in that source, or DOI.
Reports on the deduplication performed.
}
