% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{deduplicate_per_unit}
\alias{deduplicate_per_unit}
\title{Deduplicate a dataframe for each level org_unit}
\usage{
deduplicate_per_unit(df)
}
\description{
Resulting dataframe ensures no duplication exists within
organization units, but duplication remains between.
}
