% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_doaj.R
\name{doaj_api}
\alias{doaj_api}
\title{Mining the DOAJ API}
\usage{
doaj_api(issn)
}
\arguments{
\item{issn}{ISSN for journal that needs to be checked}
}
\value{
list with DOAJ results
}
\description{
This function uses an issn to mine the
DOAJ API (at doaj.org/api/v2/).
The entry for this ISSN in the DOAJ is returned.
}
