% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{full_report}
\alias{full_report}
\title{Generate a full report}
\usage{
full_report(df, name = "all")
}
\description{
Function to generate both a dataset and corresponding
figure, reporting on the classification found
in the data. Requires a name to save the report and figure.
}
