% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{get_custom}
\alias{get_custom}
\title{Read the custom file}
\usage{
get_custom(path)
}
\arguments{
\item{path}{Location of file with custom labels}
}
\value{
list with the contents of the custom label file
}
\description{
Read the custom file
}
