% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{get_doaj}
\alias{get_doaj}
\title{Get DOAJ data from indicated source}
\usage{
get_doaj(df, source)
}
\arguments{
\item{df}{data frame with an ISSN column}

\item{source}{either "api" or the path of a saved DOAJ result}
}
\value{
data frame with DOAJ results
}
\description{
This function either kickstarts mining the DOAJ API using the
issn columnn in a provided data frame, or loads saved DOAJ data.
}
