% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_upw.R
\name{get_upw}
\alias{get_upw}
\title{Get Unpaywall data from indicated source}
\usage{
get_upw(df, source = "api", email = "", save_results = F)
}
\arguments{
\item{df}{data frame with a doi column}

\item{source}{either "api" or the path of a saved unpaywall result}

\item{email}{email address of user (required by Unpaywall)}

\item{save_results}{Do you want to save the resulting data frame?}
}
\value{
data frame with unpaywall results
}
\description{
This function either kickstarts mining the Unpaywall API using the
doi columnn in a provided data frame, or loads saved Unpaywall data.
}
