% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vsnu.R
\name{get_vsnu}
\alias{get_vsnu}
\title{Read and clean VSNU data}
\usage{
get_vsnu(path)
}
\arguments{
\item{path}{path of VSNU data}
}
\value{
data frame with VSNU data
}
\description{
Read and clean VSNU data
}
