% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{individual_reports}
\alias{individual_reports}
\title{Generate many individual reports}
\usage{
individual_reports(path_report)
}
\description{
The custom-made reporting sheet is used to
write individual reports and figures for
each set of units in the reporting sheet.
}
