% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{number_to_issn}
\alias{number_to_issn}
\title{Turn string (length 8) into ISSN}
\usage{
number_to_issn(number)
}
\arguments{
\item{number}{string of length 8 (often a number)}
}
\value{
ISSN-formatted string.
}
\description{
An ISSN is formatted as NNNN-NNNX, where N is a number, and X can be any number or character.
If ISSNs are formatted differently, this can give errors when using a public API, so they can be
reformatted with this function.
A string with a different length will return NA.
}
