% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{open_clean}
\alias{open_clean}
\title{Open and clean up source data}
\usage{
open_clean(col_config, template)
}
\arguments{
\item{col_config}{vector with the original column names, sorted by a standard}

\item{template}{vector with the description of standard columns (this must be the File_info column in the excel template)}
}
\value{
cleaned data frame
}
\description{
This function opens a source file, then changes column headers to a standard format,
removes any superfluous columns (if indicated in the excel template), and cleans DOI
and ISSN columns using `clean_doi` and `clean_issn` functions.
}
