% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_df.R
\name{open_df}
\alias{open_df}
\title{Open a single file to}
\usage{
open_df(file, dir = "")
}
\arguments{
\item{file}{The (entire!) path to the (filled out) excel template file.}

\item{dir}{The path to the folder in which all data content files are located.}
}
\value{
a data frame combining all file content
}
\description{
This function can be used in the package OAmonitor to
open all files on which information is filled out in an
excel template.
The files will have to be stored in the data folder
and formatted either as tsv, csv, xls, or xlsx.
}
