% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{open_everything}
\alias{open_everything}
\title{Open all files and return a dataframe}
\usage{
open_everything(allfiles)
}
\arguments{
\item{allfiles}{The content of the excel template, read with \code{read_excel()}.}
}
\value{
a data frame combining all file content
}
\description{
This function can be used in the package OAmonitor to
open all files on which information is filled out in an
excel template.
The files will have to be stored in the data folder
and formatted either as tsv, csv, xls, or xlsx.
}
