% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_doaj.R
\name{process_doaj}
\alias{process_doaj}
\title{Process DOAJ API result}
\usage{
process_doaj(df)
}
\arguments{
\item{df}{the data frame resulting from a DOAJ API mining}
}
\value{
cleaned up data frame
}
\description{
Specifically used to process the data frame that results from a doaj mining query.
pISSN and eISSN numbers are in separate columns;
the dataframe needs to be transformed into long format,
with issns in one column and issn types in another column
}
