% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{read_ext}
\alias{read_ext}
\title{Determines format and reads in tabular data}
\usage{
read_ext(fn, ext = "", dir = "data/")
}
\arguments{
\item{fn}{the filename}

\item{ext}{the file extension (may also be provided by filename). Can be "csv", "tsv", "xls", or "xlsx".}

\item{dir}{location of the file}
}
\value{
a data frame
}
\description{
Using the filename, this function uses tidyverse read
functions to open a data file.
}
