% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{report_to_alluvial}
\alias{report_to_alluvial}
\title{Turn classification data into an alluvial diagram}
\usage{
report_to_alluvial(df, title = "all", save = F)
}
\arguments{
\item{df}{The dataframe with classification label (OA_label; result of `classify_oa`)}

\item{title}{the name of the reporting unit}

\item{save}{show the image (F) or save it (in `figures/`)}
}
\description{
Turn a summary report into an alluvial diagram clarifying OA strategy.
}
