% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{report_to_dataframe}
\alias{report_to_dataframe}
\title{Generate a report}
\usage{
report_to_dataframe(df, title = "all", save = F)
}
\arguments{
\item{df}{The dataframe with classification label (OA_label; result of `classify_oa`)}

\item{title}{the name of the reporting unit}

\item{save}{show the image (F) or save it (in `output`)}
}
\value{
summary of the results in a dataframe
}
\description{
Turn a dataframe with classification into a report summarizing the different kinds of
access to publications.
}
