% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{upw_api}
\alias{upw_api}
\title{collecting DOI results from Unpaywall using their REST API}
\usage{
upw_api(doi, email = email_address)
}
\description{
collecting DOI results from Unpaywall using their REST API
}
