% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{upw_pipeline}
\alias{upw_pipeline}
\title{Get Unpaywall data from indicated source}
\usage{
upw_pipeline(df, source = "api", email)
}
\arguments{
\item{df}{data frame with a doi column}

\item{source}{either "api" or the path of a saved unpaywall result}

\item{email}{email address of user (required by Unpaywall)}
}
\value{
data frame with unpaywall results
}
\description{
This function either kickstarts mining the Unpaywall API using the
doi columnn in a provided data frame, or loads saved Unpaywall data.
}
