% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{api_to_df}
\alias{api_to_df}
\title{Using an API mining  function, query all unique
entries in a column, and return a data frame
with their results.}
\usage{
api_to_df(df, which_info)
}
\arguments{
\item{which_info}{What api will be mined? Either "doaj" or "upw" (unpaywall).}
}
\description{
Using an API mining  function, query all unique
entries in a column, and return a data frame
with their results.
}
