% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{doaj_api}
\alias{doaj_api}
\title{This function uses an issn to mine the
DOAJ API (at doaj.org/api/v1/).
The entry for this ISSN in the DOAJ is returned.}
\usage{
doaj_api(issn)
}
\description{
This function uses an issn to mine the
DOAJ API (at doaj.org/api/v1/).
The entry for this ISSN in the DOAJ is returned.
}
