% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{full_report}
\alias{full_report}
\title{Generate a full report}
\usage{
full_report(df, title = "all")
}
\arguments{
\item{df}{The dataframe with classification label (OA_label; result of `classify_oa`)}

\item{title}{the name of the reporting unit}
}
\description{
Function that runs all individual reporting functions, and generates
a report dataframe in `data/output`, and images in `figures`.
For individual reports that directly return results: use `report_to_dataframe`,
`report_to_image`, or `report_to_alluvial`.
}
