% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{hoop_report}
\alias{hoop_report}
\title{Report for HOOP areas}
\usage{
hoop_report(df)
}
\description{
Go over HOOP areas that were filled out with existing organization units
and generate reports on the areas overall.
}
