% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{individual_reports}
\alias{individual_reports}
\title{Generate many individual reports}
\usage{
individual_reports(df, path_report)
}
\arguments{
\item{df}{The dataframe with classification label (OA_label; result of `classify_oa`)}

\item{path_report}{Path to the custom excel file that contains reporting information}
}
\description{
The custom-made reporting sheet is used to
write individual reports and figures for
each set of units in the reporting sheet.
}
