% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{process_doaj}
\alias{process_doaj}
\title{Specifically used to process the data frame
that results from a doaj mining query.
ISSN numbers are in a nested format, 
they need to be un-nested.
For the ease of future processing, the ISSN
columns are renamed.}
\usage{
process_doaj(df)
}
\description{
Specifically used to process the data frame
that results from a doaj mining query.
ISSN numbers are in a nested format, 
they need to be un-nested.
For the ease of future processing, the ISSN
columns are renamed.
}
