% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{save_df}
\alias{save_df}
\title{Save dataframes in data/clean}
\usage{
save_df(df, which_info)
}
\arguments{
\item{df}{data frame to save}

\item{which_info}{either 'doaj' for DOAJ mining results, 'upw', for Unpaywall, 'all', for all data, or 'check' for publications that need to be verified.}
}
\description{
Save three kinds of clean data: results from APIs and total data after classification.
}
